standardize streaming of intermediate output
